Refactor settings, remove color table
The settings handling is replaced with a simpler template
construction that avoids code duplication.

The new setting class allows for specification if setting
should be put in global config or just in workspace config

The color table enabled the default colors to be altered, this
is an old construct that is now removed.

Fix ticket 204
Row Clipping stays enabled after LS shutdown