Fix first bookmark assert
On recent Qt debug versions there is an assert in-case
an item is added to an empty parent with index different
from 0. The LS error was that a pointer to a previous
bookmark is always provided, and if no bookmark was provided
then the bookmark root was provided as previous.

When cleaning the DB the LUT was as well emptied, however
when filters are supposed to remain then it should not be
cleaned.